pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "weather-api"
version = "0.1.0"
description = "Weather forecast API built with FastAPI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-httpx>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "prometheus-client>=0.24.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.7.0",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.40.0",
    "slowapi>=0.1.9",
    "redis>=5.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/weather_api"]

[dependency-groups]
dev = [
    "fakeredis>=2.20.0",
    "httpx>=0.28.1",
    "locust>=2.43.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "respx>=0.22.0",
    "ruff>=0.14.14",
]

[tool.mypy]
strict = true
python_version = "3.12"
mypy_path = "src"
packages = ["weather_api"]

[tool.ruff]
src = ["src"]
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "T10",    # debugger statements
    "C901",   # McCabe complexity
    "RUF",    # Ruff-specific rules
    "ASYNC",  # async anti-patterns
    "S",      # bandit security checks
    "PTH",    # pathlib enforcement
]
ignore = [
    "S101",   # Allow assert in tests
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"tests/performance/*" = ["S311"]  # Allow random.choice for load testing

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["weather_api"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/observability/logging.py",   # Configuration-heavy, tested via integration
    "*/observability/tracing.py",   # Configuration-heavy, tested via integration
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
fail_under = 80
show_missing = true
